+++ com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/HDRenderPipeline.cs
@@ -1729,6 +1729,11 @@ namespace UnityEngine.Rendering.HighDefinition
                         // Ideally once each frame
                         {
                             var cmd = CommandBufferPool.Get("");
+                            using (new ProfilingScope(cmd, ProfilingSampler.Get(HDProfileId.BuildMaginals)))
+                            {
+                                // Generate Marginals texture for importance sampling
+                                ImportanceSamplers.Update(cmd);
+                            }
                             // Release a RTHandle when their lifetime counter scheduled became 0
                             RTHandleDeleter.Update();
 
@@ -2079,7 +2084,6 @@ namespace UnityEngine.Rendering.HighDefinition
             }
             else
             {
-
                 // When debug is enabled we need to clear otherwise we may see non-shadows areas with stale values.
                 if (hdCamera.frameSettings.IsEnabled(FrameSettingsField.ContactShadows) && m_CurrentDebugDisplaySettings.data.fullScreenDebugMode == FullScreenDebugMode.ContactShadows)
                 {